Make collision detection muuuuuuch better
When the game was run at a high enough speed previously,
it was possible for the ball to be entirely to the left
of a paddle in one frame and then entirely to the right
by the next one. Our collision detection algorithm didn't
detect this so it was possible to fire the ball straight
through a paddle without ever making contact with it.

This commit replaces that collision detection algorithm
with one that compares the vectors each of the ball's four
corners will follow during the next tick to the four line
segments defined by the paddle's edges. This should give us
ridiculously accurate collision detection and make it
utterly impossible to ever drop a ball through a paddle.

One down side of this is we never actually see the ball
touch the paddle since the frame before it would have we
are reflecting it away. Maybe we can set some kind of
threshold where if the ball is more than 75% of its vector's
magnitude away from the paddle, we pretend it was 100%, and render
a frame with the two right next to each other at the expense of
of a little simplicity.